+++ Changes.md
@@ -13,6 +13,7 @@ Fixes
 - CodeWidget :
   - Removed unwanted `\n` prefix in lines passed to `Completers`.
   - Prevented spurious emission of `editingFinishedSignal()` when showing the completions menu.
+  - Prevented the completion menu from popping up inappropriately when the highlighter was changed.
 
 API
 ---